Increased size of indices to 32bit

+++ include/render/Model.h
@@ -18,7 +18,7 @@ class Model {
 public:
 	// Type aliases
 	using VertexType = TextureVertex;
-	using IndexType = Uint16;
+	using IndexType = Uint32;
 	using MeshType = Mesh<VertexType, IndexType>;
 
 private:
